requirements.txt
flask
flasgger
flask-login
sqlalchemy
psycopg2-binary
pyyaml
pytest
requests
bandit